Cherry pick of 61484cffb572bbc7edf5039ecfae2c49e7f4a305 failed
stdout:
CONFLICT (modify/delete): payment_mercadopago/models/payment.py deleted in HEAD and modified in 61484cf... [FIX] Evaluate if request is available when define the CVV_token. If not available assume that the token is valid.. Version 61484cf... [FIX] Evaluate if request is available when define the CVV_token. If not available assume that the token is valid. of payment_mercadopago/models/payment.py left in tree.
Auto-merging payment_mercadopago/__manifest__.py
CONFLICT (content): Merge conflict in payment_mercadopago/__manifest__.py

stderr:
21:12:41.608404 git.c:439               trace: built-in: git cherry-pick 61484cffb572bbc7edf5039ecfae2c49e7f4a305
error: could not apply 61484cf... [FIX] Evaluate if request is available when define the CVV_token. If not available assume that the token is valid.
hint: after resolving the conflicts, mark the corrected paths
hint: with 'git add <paths>' or 'git rm <paths>'
hint: and commit the result with 'git commit'
----------
status: